# sentinel-ai/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SentinelAI API
  version: 0.1.0
  description: |
    Implemented: the security-engine scan contract (`/v1/scan`, internal) and the gateway routes for security scan/check,
    AI chat/generate, file scanning (`/v1/files/scan`), events, usage and policies (API-key auth), login/JWT, user and
    team management (`/v1/users`, `/v1/invitations`, `/v1/teams`) and per-organization provider credentials
    (`/v1/providers`). Routes marked (planned) are specified ahead of implementation.

    Management routes (users, teams, invitations, providers, API keys) require a USER session (JWT): an API key is never
    accepted for them, so a leaked key cannot create users, change roles or store provider credentials.
    Wire format is snake_case JSON. Responses never contain matched sensitive values.
servers:
  - url: http://localhost:4000
    description: API gateway
  - url: http://localhost:8001
    description: Security engine (implemented, internal only)
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/red-team/rounds:
    get: { summary: "Red-team rounds, newest first (implemented; events:read)", parameters: [{ name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }], responses: { "200": { description: "{ rounds }" } } }
    post:
      summary: |
        Record one red-team round (implemented; evaluation:run). Counts per category and tier plus at most 40 synthetic,
        sanitized, truncated examples; append-only (409 if the round number exists). Written by scripts/security/red_team.py.
      responses: { "201": { description: recorded }, "409": { description: round_exists }, "422": { description: invalid (blocked + slipped must equal attacks) } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKeyAuth: { type: apiKey, in: header, name: X-Sentinel-Api-Key }
    internalToken: { type: apiKey, in: header, name: X-Internal-Token }
...